Consolidate into one test file
